+++ basana/core/dispatcher.py
@@ -242,7 +242,8 @@ class EventDispatcher(metaclass=abc.ABCMeta):
                     tg.create_task(producer.main())
                 tg.create_task(self._dispatch_loop())
         except asyncio.CancelledError:
-            pass
+            if not self.stopped:
+                raise
         finally:
             # Cancel any pending task in the event pool.
             self._handlers_task_pool.cancel()
@@ -323,9 +324,13 @@ class BacktestingDispatcher(EventDispatcher):
         with logs.backtesting_log_mode(self):
             await super().run(stop_signals=stop_signals)
 
+    @property
+    def now_available(self) -> bool:
+        return self._last_dt is not None
+
     def now(self) -> datetime.datetime:
         if self._last_dt is None:
-            raise errors.Error("No events processed yet")
+            raise errors.Error("Can't calculate current datetime since no events were processed")
         return self._last_dt
 
     def _set_now(self, now: datetime.datetime):