+++ basana/core/websockets.py
@@ -69,6 +69,9 @@ class WebSocketClient(event.Producer, metaclass=abc.ABCMeta):
     def schedule_reconnection(self):
         self._reconnect_request.set()
 
+    def schedule_resubscription(self, channels: List[str]):
+        self._pending_subscriptions.update(channels)
+
     async def on_error(self, error: Any):
         logger.error(logs.StructuredMessage("Error", src=self, error=error))
 
@@ -130,7 +133,7 @@ class WebSocketClient(event.Producer, metaclass=abc.ABCMeta):
             if not handled:
                 await self.on_unknown_message(message)
 
-        # If the message loop finished we need to gracefully stop the other tasks.
+        # If the message loop finished we need to notify the other tasks so they can finish as well.
         self._subscribe_request.set()
         self._reconnect_request.set()
 
@@ -138,6 +141,7 @@ class WebSocketClient(event.Producer, metaclass=abc.ABCMeta):
         # Will exit when reconnection is requested or when its canceled.
         await self._reconnect_request.wait()
         self._reconnect_request.clear()
+        # If the client is already closed then there is nothing left to do.
         if not ws_cli.closed:
             await ws_cli.close()