+++ samples/binance_websockets.py
@@ -44,10 +44,21 @@ async def on_trade_event(trade_event: binance_exchange.TradeEvent):
     )
 
 
+async def on_order_event(event):
+    logging.info(
+        "Order event: id=%s status=%s amount_filled=%s fees=%s",
+        event.order_update.id, event.order_update.status, event.order_update.amount_filled, event.order_update.fees
+    )
+
+
 async def main():
     logging.basicConfig(level=logging.INFO, format="[%(asctime)s %(levelname)s] %(message)s")
     event_dispatcher = bs.realtime_dispatcher()
-    exchange = binance_exchange.Exchange(event_dispatcher)
+    exchange = binance_exchange.Exchange(
+        event_dispatcher,
+        # api_key="YOUR_API_KEY",
+        # api_secret="YOUR_API_SECRET"
+    )
 
     pairs = [
         bs.Pair("BTC", "USDT"),
@@ -57,6 +68,13 @@ async def main():
         exchange.subscribe_to_bar_events(pair, "1m", on_bar_event)
         exchange.subscribe_to_order_book_events(pair, on_order_book_event)
         exchange.subscribe_to_trade_events(pair, on_trade_event)
+
+    # Uncomment the following lines if you want to subscribe to order events. This requires the API key and secret to
+    # be set.
+    # exchange.spot_account.subscribe_to_order_events(on_order_event)
+    # exchange.cross_margin_account.subscribe_to_order_events(on_order_event)
+    # for pair in pairs:
+    #     exchange.isolated_margin_account.subscribe_to_order_events(pair, on_order_event)
 
     await event_dispatcher.run()