+++ tests/test_binance_order_book.py
@@ -160,7 +160,7 @@ def test_websocket_ok(realtime_dispatcher):
         assert message["method"] == "SUBSCRIBE"
         await websocket.send(json.dumps({"result": None, "id": message["id"]}))
 
-        while True:
+        while websocket.state == websockets.protocol.State.OPEN:
             await websocket.send(json.dumps({"stream": "btcusdt@depth10", "data": ORDER_BOOK}))
             await websocket.send(json.dumps({"stream": "btcusdt@depth20", "data": ORDER_BOOK}))
             await asyncio.sleep(0.1)